+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.1] – 2025-09-26
+
+### Changed
+- Hyperliquid market order price deviation
+- Hyperliquid spot reduce only flag
+- Hyperliquid retry get order
+
 ## [1.1.0] – 2025-09-24
 
 ### Added